refactor(workspace): collapse materializer/sqlite/types.ts; rename sqlite.ts to core.ts
Mechanical cleanup ahead of per-backend materializer attach functions. No
behavior change; tests pass (485/485).

- sqlite.ts -> core.ts. The materializer body is about to be shared across
  multiple backend-named attach functions, so "sqlite.ts" misnames the role.
  "core" describes what the file actually owns: the shared materializer.
- sqlite.test.ts -> core.test.ts, matching the renamed source.
- types.ts dissolved. It was collecting two unrelated audiences under a
  meaningless name: the internal SQL executor contract (MirrorDatabase,
  MirrorStatement) and the public search response shapes (SearchOptions,
  SearchResult). Internal types now live at the top of core.ts marked
  @internal; public search types live in fts.ts (the natural home; FTS
  already owns the search behavior).
- The materializer barrel (materializer/sqlite/index.ts) stops re-exporting
  MirrorDatabase and MirrorStatement. They are no longer public surface;
  per-backend attach functions will own the adapter from a native client
  to that contract.
