+++ packages/workspace/src/document/materializer/sqlite/fts.ts
@@ -9,9 +9,43 @@
  */
 
 import type { Logger } from 'wellcrafted/logger';
+import { SqliteMaterializerError, type MirrorDatabase } from './core.js';
 import { quoteIdentifier } from './ddl.js';
-import { SqliteMaterializerError } from './sqlite.js';
-import type { MirrorDatabase, SearchOptions, SearchResult } from './types.js';
+
+// ════════════════════════════════════════════════════════════════════════════
+// PUBLIC SEARCH TYPES
+// ════════════════════════════════════════════════════════════════════════════
+
+/**
+ * Optional arguments for FTS5 searches.
+ *
+ * Use this when you want to cap result count or choose which indexed column is
+ * used for snippets in the search response.
+ */
+export type SearchOptions = {
+	/** Maximum number of matches to return. */
+	limit?: number;
+
+	/** Column name used to generate the snippet text. */
+	snippetColumn?: string;
+};
+
+/**
+ * One full-text search result returned by the materializer.
+ *
+ * `id` points back to the materialized row, `snippet` is display-ready text, and
+ * `rank` is the database-provided relevance score.
+ */
+export type SearchResult = {
+	/** ID of the materialized row that matched the query. */
+	id: string;
+
+	/** Snippet generated from indexed text content. */
+	snippet: string;
+
+	/** Relevance score returned by the FTS query. */
+	rank: number;
+};
 
 // ════════════════════════════════════════════════════════════════════════════
 // FTS SETUP
