+++ packages/workspace/src/document/open-sqlite-reader.test.ts
@@ -12,7 +12,7 @@ import { join } from 'node:path';
 import { createLogger } from 'wellcrafted/logger';
 import * as Y from 'yjs';
 import { attachTables, column, defineTable } from '../index.js';
-import { attachSqliteMaterializer } from './materializer/sqlite/sqlite.js';
+import { attachSqliteMaterializer } from './materializer/sqlite/core.js';
 import { openSqliteReader } from './open-sqlite-reader.js';
 import { openWriterSqlite } from './sqlite-writer.js';
